requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
anyio==4.4.0
appnope==0.1.4
asttokens==2.4.1
attrs==23.2.0
blinker==1.8.2
blis==0.7.11
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
comm==0.2.2
confection==0.1.5
cymem==2.0.8
debugpy==1.8.2
decorator==5.1.1
diskcache==5.6.3
distro==1.9.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
executing==2.0.1
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
gguf==0.6.0
gitdb==4.0.11
GitPython==3.1.43
groq==0.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.23.4
idna==3.7
ipykernel==6.29.4
ipython==8.26.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
langchain==0.2.6
langchain-core==0.2.10
langchain-text-splitters==0.2.2
langcodes==3.4.0
langsmith==0.1.82
language_data==1.2.0
llama_cpp_python==0.2.79
lxml==5.2.2
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
nest-asyncio==1.6.0
networkx==3.3
numpy==1.26.4
openai==1.35.7
orjson==3.10.5
packaging==24.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
pinecone-client==4.1.1
pinecone-plugin-interface==0.0.7
platformdirs==4.2.2
preshed==3.0.9
prompt_toolkit==3.0.47
protobuf==5.27.2
psutil==6.0.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.3
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
scikit-learn==1.5.0
scipy==1.14.0
sentence-transformers==3.0.1
setuptools==70.1.1
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.31
srsly==2.4.8
stack-data==0.6.3
streamlit==1.36.0
sympy==1.12.1
tenacity==8.4.2
thinc==8.2.5
threadpoolctl==3.5.0
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
wrapt==1.16.0
yarl==1.9.4
